requirements.txt
fastapi
uvicorn
pandas
matplotlib
plotly
psycopg2-binary
seaborn
jinja2